refactor: 상속하는 파일들 패키지로 분리
