+++ src/main/java/roomescape/ExceptionController.java
@@ -6,6 +6,10 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import roomescape.exception.AuthenticationException;
+import roomescape.exception.AuthorizationException;
+import roomescape.exception.ConflictException;
+import roomescape.exception.NotFoundException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
