+++ src/main/java/roomescape/exception/AuthenticationException.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.exception;
 
 public class AuthenticationException extends RuntimeException {
 }
