+++ src/main/java/roomescape/exception/ConflictException.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.exception;
 
 public class ConflictException extends RuntimeException {
     public ConflictException(String message) {
