+++ src/main/java/roomescape/exception/NotFoundException.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.exception;
 
 public class NotFoundException extends RuntimeException {
     public NotFoundException(String message) {
