+++ src/main/java/roomescape/reservation/ReservationService.java
@@ -5,9 +5,9 @@ import org.slf4j.LoggerFactory;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.AuthenticationException;
-import roomescape.ConflictException;
-import roomescape.NotFoundException;
+import roomescape.exception.AuthenticationException;
+import roomescape.exception.ConflictException;
+import roomescape.exception.NotFoundException;
 import roomescape.auth.LoginMemberInfo;
 import roomescape.member.Member;
 import roomescape.member.MemberDao;
