+++ src/main/java/roomescape/theme/ThemeService.java
@@ -2,7 +2,7 @@ package roomescape.theme;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.NotFoundException;
+import roomescape.exception.NotFoundException;
 
 import java.util.List;
