+++ src/main/java/roomescape/time/TimeService.java
@@ -2,7 +2,7 @@ package roomescape.time;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.NotFoundException;
+import roomescape.exception.NotFoundException;
 import roomescape.reservation.Reservation;
 import roomescape.reservation.ReservationDao;
